requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
rank_bm25>=0.2.2
jieba>=0.42.1
